refactor(cli): home the dev-only command surfaces under modules/harness/dev
`src/cli/index.ts` registers `chat`, `run`, and `profile` only when
`devCommandsEnabled()` is true, so a release build carries none of them.
The tree said nothing: the three sat beside the product modules, and a
reader met them as peers of the harness runtime.

They now live under `modules/harness/dev/`, so the tree states the
channel where a reader meets it first, beside the registration gate that
already states it. The dependency runs one way — `dev/` can import
product code, and no product file imports `dev/`. The three lazy
`import()` calls in the gated registration block are the one crossing,
and a release build never evaluates them.

`chat_printer.ts` split on that line. Five readers stay: the TUI
conversation reducer and the REPL printer both call them, and the two
surfaces must narrate one event stream identically. The printer folds
INTO `dev/chat.ts` rather than into a file of its own, because
`createChatPrinter` has one caller and a single-caller helper stays with
its caller.

`eventDepth` joins `conversation.ts`, its one caller — the printer never
asked the question. It inlines the source read rather than widen the
module surface by exporting a private helper.

`dev/status.ts` holds the wait and status-pool readers. Both dev commands
call them and nothing else does: the TUI activity panel refuses them on
purpose, because the step cache records a step only when that step
returns, so its newest row names whatever finished last and never the
work in flight.

`profile.ts` is gone. No file is left with a name that describes none of
its residents.

Refs: inflexa-ai/inflexa#56
