+++ cli/src/modules/harness/dev/chat.test.ts
