+++ cli/src/modules/harness/dev/chat.ts
@@ -13,20 +13,25 @@
 // coarse stdout printer. The spec-level record is
 // `openspec/specs/chat-command/spec.md`.
 
+import { pathToFileURL } from "node:url";
+
 import { randomUUIDv7 } from "bun";
 import { intro, log, outro, spinner, text, isCancel } from "@clack/prompts";
 import { type ResultAsync } from "neverthrow";
-import { createStreamingChat, createThreadHistory, createThreadStore, type AgentSession, type DbError, type Thread } from "@inflexa-ai/harness";
+import { createStreamingChat, createThreadHistory, createThreadStore, type AgentSession, type DbError, type EmitFn, type Thread } from "@inflexa-ai/harness";
+import type { OpenableEntry, OpenTarget, PresentationBody } from "../../../types/session.ts";
 
-import { describeCause } from "../../lib/cause.ts";
-import { fail } from "../../lib/cli.ts";
-import { shutdown } from "../../lib/shutdown.ts";
-import { claimAnalysisOrFail, resolveSingleAnalysis, type ContextFlags } from "../analysis/context.ts";
-import { resolveHarnessConfig } from "./config.ts";
-import { createChatPrinter, type ChatSink } from "./chat_printer.ts";
-import { describeBootError, ensureSandboxImage } from "./profile.ts";
-import { bootHarnessRuntime, type HarnessRuntime } from "./runtime.ts";
-import { buildChatSession, runChatTurn } from "./turn.ts";
+import { describeCause } from "../../../lib/cause.ts";
+import { fail } from "../../../lib/cli.ts";
+import { shutdown } from "../../../lib/shutdown.ts";
+import { claimAnalysisOrFail, resolveSingleAnalysis, type ContextFlags } from "../../analysis/context.ts";
+import { resolveHarnessConfig } from "../config.ts";
+import { ensureSandboxImage } from "../../libs/pull.ts";
+import { materializeTarget, readFileReference, readPresentation, readReportPreview, readReportPreviewFailed } from "../artifact_open.ts";
+import { isSubAgentEvent, readAskPart, readPlanCard, readRunCard, subAgentActivityLabel } from "../chat_printer.ts";
+import { planToDag } from "../plan_dag.ts";
+import { bootHarnessRuntime, describeBootError, type HarnessRuntime } from "../runtime.ts";
+import { buildChatSession, runChatTurn } from "../turn.ts";
 
 /** The `empty`-context hint specific to `inflexa chat`. */
 const CHAT_EMPTY_HINT = "No analysis here. Run `inflexa` to start one, add inputs, then `inflexa chat`.";
@@ -317,4 +322,284 @@ async function runTurn(
     } finally {
         process.removeListener("SIGINT", onSigint);
     }
+}
+
+// ── The emit sink ────────────────────────────────────────────────────────────────────────────────
+//
+// Renders one in-process `EmitFn` stream to a plain-text terminal. Deliberately coarse: this is the
+// dev surface, not the TUI renderer. It lives in this file rather than beside it because
+// `createChatPrinter` has exactly one caller — the REPL above — and a single-caller helper stays with
+// its caller (`cli/CLAUDE.md`). The event READERS it calls are shared with the TUI, so those stay in
+// `modules/harness/chat_printer.ts`.
+//
+// Three rules are load-bearing and each maps to a chat-command spec requirement:
+//
+//   1. COPY-ON-RECEIVE. In-process `emit` shares mutable references with the agent loop (the same
+//      hazard the TUI's clone-before-store rule guards). Every branch extracts the strings, ids, and
+//      statuses it renders at receipt and NEVER retains the received event or its `data` object.
+//      Printing is synchronous inside `emit`, so a caller that mutates a part after emitting it
+//      cannot change what was already written.
+//   2. TOP-LEVEL ONLY. Events whose `source.callPath` is deeper than the top-level agent (sub-agent
+//      traffic: planner, literature reviewer) are routed under the tool call they run inside, never
+//      to the transcript root.
+//   3. ACCUMULATE, RENDER COARSELY. Text deltas are written as received; the terminal itself is the
+//      accumulator. Tool activity prints a one-line chip on start and its outcome on finish.
+//      `data-plan`/`data-run-card` render their embedded content; every other conversation part
+//      prints a one-line tagged mention so the surface OBSERVES unknown traffic rather than hiding it.
+//
+// stdout carries the conversation; stderr carries diagnostics — the sink splits them so a caller can
+// pipe the transcript cleanly. Output is plain ASCII: the `GLYPHS` registry is a `src/tui/` rule.
+
+/**
+ * Where the printer writes. Injected so the unit tests drive a pure recording
+ * sink; production wires `out → process.stdout.write` and `errLine → console.error`.
+ */
+export type ChatSink = {
+    /** Conversation output — written verbatim, no trailing newline added (deltas accumulate). */
+    readonly out: (s: string) => void;
+    /** One diagnostic line to stderr (a newline is the sink's concern). */
+    readonly errLine: (s: string) => void;
+};
+
+/**
+ * The small per-turn API the chat REPL drives. `emit` is the `EmitFn` handed to
+ * `runAgent` — the command also routes the streaming provider wrapper's per-token
+ * `onText` callback through it as `text-delta` events (see chat.ts), so deltas
+ * and loop/tool events share one sink and one set of rules. `finishTurn` flushes
+ * and resets per-turn state (dangling tool chips, the streamed-text flag).
+ */
+export type ChatPrinter = {
+    /**
+     * The `EmitFn` sink handed to `runAgent` (and fed the streaming provider's
+     * text deltas). Drops sub-agent traffic, renders each event category coarsely,
+     * and never retains a received object (copy-on-receive).
+     */
+    readonly emit: EmitFn;
+    /**
+     * Close out the turn. `fallbackText` is the turn's final assistant text
+     * (from `finalText(runAgent result)`): printed only when the turn streamed
+     * no `text-delta`s — the deltas and the final text are the SAME content, so
+     * this both prevents the double print on a streamed turn and keeps a
+     * delta-less turn (or one run without the streaming wrapper) from rendering
+     * nothing.
+     */
+    readonly finishTurn: (fallbackText?: string) => void;
+};
+
+/** ms as a compact human string for the tool-chip completion line. */
+function formatMs(ms: number): string {
+    return ms < 1000 ? `${ms}ms` : `${(ms / 1000).toFixed(1)}s`;
+}
+
+/** Build an OSC 8 hyperlink whose VISIBLE text is `text` and whose target is `uri` — degrades to plain `text` on terminals without link support. */
+function hyperlink(uri: string, text: string): string {
+    return `\x1b]8;;${uri}\x07${text}\x1b]8;;\x07`;
+}
+
+/** Render a text-shaped presentation table as aligned monospace columns (the REPL's plain-text table form). */
+function formatTable(headers: string[], rows: string[][]): string {
+    const widths = headers.map((h, ci) => Math.max(h.length, ...rows.map((r) => (r[ci] ?? "").length)));
+    const line = (cells: string[]): string => `    ${cells.map((c, ci) => (c ?? "").padEnd(widths[ci] ?? 0)).join("  ")}`.trimEnd();
+    return [line(headers), line(widths.map((w) => "-".repeat(w))), ...rows.map((r) => line(headers.map((_, ci) => r[ci] ?? "")))].join("\n");
+}
+
+/**
+ * How the printer resolves an openable entry to the absolute path it links to (materializing `echart`/`svg`
+ * into the workspace's `presentations/` directory). Injectable so the printer's openable rendering is
+ * unit-testable without a booted workspace; production omits it and gets the real {@link materializeTarget}.
+ */
+export type PrinterOptions = {
+    /** The analysis whose workspace root resolves openable references. */
+    readonly analysisId?: string;
+    /** Resolve an entry's target to an absolute path (materializing when needed), or `null` when unavailable. */
+    readonly resolvePath?: (analysisId: string, target: OpenTarget) => string | null;
+};
+
+/**
+ * Build a chat printer over `sink`. Holds only per-turn primitive state (the
+ * streamed-text flag and open tool chips keyed by id → name+start-time) — never
+ * a received object — so copy-on-receive holds by construction.
+ */
+export function createChatPrinter(sink: ChatSink, options: PrinterOptions = {}): ChatPrinter {
+    const analysisId = options.analysisId ?? "";
+    const resolvePath =
+        options.resolvePath ??
+        ((aid: string, target: OpenTarget): string | null =>
+            materializeTarget(aid, target).match(
+                (path) => path,
+                () => null,
+            ));
+    let streamedText = false;
+    // toolUseId → the primitives needed to close its chip. Storing the extracted
+    // name (a string copy) and a timestamp, never the event, keeps copy-on-receive.
+    const openTools = new Map<string, { name: string; startedAt: number }>();
+
+    const emit: EmitFn = (event) => {
+        // Rule 2: sub-agent traffic (planner, literature reviewer) never becomes a
+        // TRANSCRIPT entry — its iterations and tool calls are numerous, and emitting
+        // them at the root would bury the conversation. But dropping it outright made a
+        // long tool call indistinguishable from a wedged one, so it is now ROUTED
+        // instead: a subordinate line under the tool call it is running inside. The TUI
+        // adapter shares this predicate and does the same thing with its tool block.
+        if (isSubAgentEvent(event)) {
+            const label = subAgentActivityLabel(event);
+            // Only while a tool is actually open: a sub-agent event outside any tool
+            // call has nothing to be subordinate TO, and printing it at the root is the
+            // burial this rule exists to prevent.
+            if (label && openTools.size > 0) sink.out(`    ${label}\n`);
+            return;
+        }
+
+        switch (event.type) {
+            case "text-delta":
+                // Rule 3: write as received; the terminal accumulates.
+                streamedText = true;
+                sink.out(event.text);
+                return;
+            case "done":
+                // Stream terminal marker — the text already rendered via deltas.
+                return;
+            case "iteration":
+                // Loop iteration boundary — orchestration, not transcript content.
+                return;
+            case "tool-started": {
+                const name = event.name;
+                // Opaque display text the harness computed from this call's input — printed, never parsed.
+                const detail = event.detail;
+                openTools.set(event.toolUseId, { name, startedAt: Date.now() });
+                sink.out(`\n  [tool] ${name}${detail === undefined ? "" : ` ${detail}`} running...\n`);
+                return;
+            }
+            case "tool-finished": {
+                const name = event.name;
+                const started = openTools.get(event.toolUseId);
+                openTools.delete(event.toolUseId);
+                const dur = started ? ` (${formatMs(Date.now() - started.startedAt)})` : "";
+                // Three outcomes get three words. `denied` is the user's own refusal of an approval, so
+                // printing it as `error` would report their decision as a fault of the tool.
+                const outcome = event.outcome === "error" ? "error" : event.outcome === "denied" ? "denied" : `done${dur}`;
+                sink.out(`  [tool] ${name} ${outcome}\n`);
+                return;
+            }
+            default: {
+                // Only `ChatDataPart` remains (its `type` is `data-${string}`).
+                renderDataPart(event.type, event.data);
+                return;
+            }
+        }
+    };
+
+    function renderDataPart(type: `data-${string}`, data: unknown): void {
+        switch (type) {
+            case "data-plan": {
+                const plan = readPlanCard(data);
+                const heading = plan.title || plan.planId;
+                sink.out(`\n  [plan] ${heading} (${plan.planId})\n`);
+                const graph =
+                    plan.steps.length > 0
+                        ? planToDag(plan.steps).match(
+                              (value) => value || null,
+                              () => null,
+                          )
+                        : null;
+                if (graph) {
+                    for (const line of graph.split("\n")) sink.out(`    ${line}\n`);
+                } else {
+                    for (const step of plan.steps) sink.out(`    - ${step.id} ${step.name} [${step.agent}]\n`);
+                }
+                return;
+            }
+            case "data-run-card": {
+                const run = readRunCard(data);
+                sink.out(`\n  [run] ${run.runId}: ${run.title} (${run.stepCount} step(s))\n`);
+                return;
+            }
+            case "data-presentation": {
+                const view = readPresentation(data);
+                if (view.shape === "inline") renderInlinePresentation(view.title, view.body);
+                else renderOpenables(view.title, [view.entry]);
+                return;
+            }
+            case "data-file-reference": {
+                const view = readFileReference(data);
+                renderOpenables(view.title, view.entries);
+                return;
+            }
+            case "data-report-preview": {
+                const view = readReportPreview(data);
+                renderOpenables(view.title, [view.entry]);
+                return;
+            }
+            case "data-report-preview-failed": {
+                const view = readReportPreviewFailed(data);
+                sink.out(`\n  [report] ${view.entry.name}: ${view.entry.caption ?? "failed"}\n`);
+                return;
+            }
+            case "data-ask": {
+                // The REPL is a write-only sink with no mid-turn input path, so it cannot answer an ask —
+                // the harness denies it by default. Still observe the approval and its outcome, one line.
+                const ask = readAskPart(data);
+                sink.out(`\n  [approval] ${ask.command} — ${ask.status}\n`);
+                return;
+            }
+            default:
+                // Rule 3: observe unknown parts, do not swallow them.
+                sink.out(`  [part:${type}]\n`);
+                return;
+        }
+    }
+
+    /** Print a text-shaped presentation inline: markdown source verbatim, code fenced, tables as aligned text. */
+    function renderInlinePresentation(title: string | undefined, body: PresentationBody): void {
+        if (title) sink.out(`\n  [show] ${title}\n`);
+        switch (body.kind) {
+            case "markdown":
+                sink.out(`${body.body}\n`);
+                return;
+            case "code":
+                sink.out("```" + body.language + "\n" + body.code + "\n```\n");
+                return;
+            case "table":
+                sink.out(`${formatTable(body.headers, body.rows)}\n`);
+                if (body.caption) sink.out(`    ${body.caption}\n`);
+                return;
+            default: {
+                const _exhaustive: never = body;
+                return _exhaustive;
+            }
+        }
+    }
+
+    /** Print openable entries: one line per entry with the resolved path as an OSC 8 `file://` link (plain path visible). */
+    function renderOpenables(title: string | undefined, entries: OpenableEntry[]): void {
+        if (title) sink.out(`\n  [show] ${title}\n`);
+        for (const entry of entries) {
+            if (entry.target.kind === "unavailable") {
+                sink.out(`    ${entry.name}: ${entry.caption ?? "unavailable"}\n`);
+                continue;
+            }
+            const path = resolvePath(analysisId, entry.target);
+            const suffix = entry.caption ? ` — ${entry.caption}` : "";
+            // The visible text stays the raw path; the link TARGET is a percent-encoded `file://` URI
+            // (via `pathToFileURL`) so spaces / `#` in the path don't truncate or mangle the OSC 8 target.
+            if (path) sink.out(`    ${entry.name}  ${hyperlink(pathToFileURL(path).href, path)}${suffix}\n`);
+            else sink.out(`    ${entry.name}  (path unavailable)${suffix}\n`);
+        }
+    }
+
+    function finishTurn(fallbackText?: string): void {
+        // Non-streaming `runAgent` path: nothing arrived as deltas, so print the
+        // final assistant text now (a streaming loop that emitted deltas skips this).
+        if (!streamedText && fallbackText && fallbackText.trim().length > 0) {
+            sink.out(fallbackText);
+        }
+        // A turn aborted mid-tool leaves a chip open — close it honestly.
+        for (const [, { name }] of openTools) sink.out(`  [tool] ${name} interrupted\n`);
+        // Separate this turn's output from the next prompt.
+        sink.out("\n");
+        streamedText = false;
+        openTools.clear();
+    }
+
+    return { emit, finishTurn };
 }
