+++ cli/src/modules/harness/dev/status.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
 
-import { describeBootError, friendlyStepLabel } from "./profile.ts";
+import { friendlyStepLabel } from "./status.ts";
 
 // Names as recorded in dbos.operation_outputs by the profile workflow — the
 // progress channel parses them, so pin the observed formats.
@@ -24,23 +24,5 @@ describe("friendlyStepLabel", () => {
 
     test("unknown step names pass through verbatim", () => {
         expect(friendlyStepLabel("DBOS.getResult")).toBe("DBOS.getResult");
-    });
-});
-
-// The sandbox_engine_unresolved arm carries a message already built against the
-// pinned runtime AND host platform at resolution time, so it must be surfaced
-// verbatim rather than re-wrapped.
-describe("describeBootError", () => {
-    test("sandbox_engine_unresolved surfaces the resolution message verbatim", () => {
-        const message =
-            "Could not resolve the Podman sandbox-engine socket — the Podman machine is not running.\n  Start it with `podman machine start`, then re-run.";
-        expect(describeBootError({ type: "sandbox_engine_unresolved", message })).toBe(message);
-    });
-
-    // A `cooling_down` cause must read as the self-recovering all-credential block it is, NOT as the
-    // generic "proxy is unreachable" the other `model_unresolved` causes render — otherwise the user
-    // chases a container that is fine.
-    test("model_unresolved cooling_down explains the proxy recovers on its own", () => {
-        expect(describeBootError({ type: "model_unresolved", cause: { type: "cooling_down" } })).toContain("recovers on its own");
     });
 });
